+++ src/events/eventStream.h
@@ -19,11 +19,13 @@
 #include "events/subscriptions/fileAttrSubscription.h"
 #include "events/subscriptions/fileLocationSubscription.h"
 #include "events/subscriptions/fileRemovalSubscription.h"
+#include "events/subscriptions/fileRenamedSubscription.h"
 #include "events/subscriptions/permissionChangedSubscription.h"
 #include "events/subscriptions/quotaSubscription.h"
 #include "events/subscriptions/readSubscription.h"
 #include "events/subscriptions/writeSubscription.h"
 #include "events/types/fileRemovalEvent.h"
+#include "events/types/fileRenamedEvent.h"
 #include "events/types/permissionChangedEvent.h"
 #include "events/types/quotaExeededEvent.h"
 #include "events/types/readEvent.h"
@@ -63,6 +65,10 @@ using FileRemovalEventStream =
 using QuotaExeededEventStream =
     EventWorker<EventCounterAggregator<EventTimeAggregator<SubscriptionHandler<
         EventHandler<EventCommunicator<QuotaExeededEvent>>>>>>;
+
+using FileRenamedEventStream =
+    EventWorker<EventCounterAggregator<EventTimeAggregator<SubscriptionHandler<
+        EventHandler<EventCommunicator<FileRenamedEvent>>>>>>;
 
 } // namespace events
 } // namespace client