+++ src/events/types/permissionChangedEvent.cc
@@ -38,7 +38,7 @@ std::string PermissionChangedEvent::toString() const
 {
     std::stringstream stream;
     stream << "type: 'PermissionChangedEvent', counter: " << m_counter
-           << ", file UUID: '" << m_fileUuid;
+           << ", file UUID: '" << m_fileUuid << "'";
     return stream.str();
 }