+++ src/fsSubscriptions.cc
@@ -75,6 +75,22 @@ void FsSubscriptions::removeFileRemovalSubscription(const std::string &fileUuid)
     }
 }
 
+void FsSubscriptions::addFileRenamedSubscription(const std::string &fileUuid)
+{
+    typename decltype(m_fileRenamedSubscriptions)::accessor acc;
+    if (m_fileRenamedSubscriptions.insert(acc, fileUuid))
+        acc->second = sendFileRenamedSubscription(fileUuid);
+}
+
+void FsSubscriptions::removeFileRenamedSubscription(const std::string &fileUuid)
+{
+    typename decltype(m_fileRenamedSubscriptions)::accessor acc;
+    if (m_fileRenamedSubscriptions.find(acc, fileUuid)) {
+        sendSubscriptionCancellation(acc->second);
+        m_fileRenamedSubscriptions.erase(acc);
+    }
+}
+
 void FsSubscriptions::addFileAttrSubscription(const std::string &fileUuid)
 {
     typename decltype(m_fileAttrSubscriptions)::accessor acc;
@@ -152,6 +168,16 @@ std::int64_t FsSubscriptions::sendQuotaSubscription()
     DLOG(INFO) << "Sending subscription for quota";
     events::QuotaSubscription clientSubscription{};
     events::QuotaSubscription serverSubscription{};
+    return m_eventManager.subscribe(
+        std::move(clientSubscription), std::move(serverSubscription));
+}
+
+std::int64_t FsSubscriptions::sendFileRenamedSubscription(
+    const std::string &fileUuid)
+{
+    DLOG(INFO) << "Sending subscription for renaming file: " << fileUuid;
+    events::FileRenamedSubscription clientSubscription{fileUuid};
+    events::FileRenamedSubscription serverSubscription{fileUuid};
     return m_eventManager.subscribe(
         std::move(clientSubscription), std::move(serverSubscription));
 }